requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
psutil
openai
beautifulsoup4
